moodVine_ai/requirements.txt
flask>=2.0.0
requests>=2.25.0
requests>=2.28.0
Pillow>=9.3.0
transformers>=4.26.0
torch>=1.13.0